requirements.txt
absl-py==2.0.0
beautifulsoup4==4.12.2
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
entrypoints==0.3
filelock==3.13.1
flake8==3.7.9
fonttools==4.44.0
fsspec==2023.10.0
ftfy==6.1.1
future==0.18.3
gdown==4.7.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
grad-cam==1.4.8
grpcio==1.59.2
idna==3.4
importlib-resources==6.1.1
install==1.3.5
isort==4.3.21
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
littleutils==0.2.2
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
mccabe==0.6.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ogb==1.3.6
opencv-python==4.8.1.78
outdated==0.2.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycodestyle==2.5.0
pyflakes==2.1.1
pymongo==4.6.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-gradcam==0.2.1
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
soupsieve==2.5
sympy==1.12
tabulate==0.9.0
tb-nightly==2.16.0a20231111
tensorboard==2.15.1
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
torch==2.1.0
torch-tb-profiler==0.4.3
torchray==1.0.0.2
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
ttach==0.0.3
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
Werkzeug==3.0.1
wilds==1.2.2
yacs==0.1.8
yapf==0.29.0
